feat(dashboards): scope email drawer to email, fill attribution drawer
The Email drawer surfaced paid and attribution content left over from the
combined Campaign Performance card, while the Attribution drawer rendered
nothing above its Event Registration section.

Move the Marketing Attribution section — multi-touch model totals and the
Revenue by Channel table — into the Attribution drawer, which now fetches
getMarketingAttribution itself since that payload is separate from the
revenueImpact input the rest of that drawer binds to. Restrict the Email
drawer's insights and recommendations to the email bucket; the paid and
attribution buckets are still computed because they feed the shared
funnel maths, but no longer render there.

Also reorders Sentiment ahead of Flywheel to match the agreed card
sequence. Education and Web are not included — neither has a data source
on EdEvolutionData yet.

LFXV2-2023
