+++ apps/lfx-one/src/app/modules/dashboards/executive-director/components/email-ctr-drawer/email-ctr-drawer.component.html
@@ -114,118 +114,6 @@
         </div>
       }
 
-      @if (attributionData().channels.length > 0) {
-        <div class="flex flex-col gap-5" data-testid="email-ctr-drawer-attribution-section">
-          <div class="flex items-center gap-2">
-            <i class="fa-light fa-chart-mixed text-indigo-500"></i>
-            <h3 class="text-base font-semibold text-gray-900">Marketing Attribution</h3>
-          </div>
-
-          <!-- Attribution Summary Stats -->
-          <div class="flex flex-wrap gap-4" data-testid="email-ctr-drawer-attribution-stats">
-            <lfx-card styleClass="flex-1 min-w-[100px]">
-              <div class="flex flex-col gap-1">
-                <span class="text-sm text-gray-500">Total Sessions</span>
-                <span class="text-2xl font-semibold text-gray-900">{{ formatCompact(attributionTotals().sessions) }}</span>
-              </div>
-            </lfx-card>
-            <lfx-card styleClass="flex-1 min-w-[100px]">
-              <div class="flex flex-col gap-1">
-                <span class="text-sm text-gray-500">First Touch</span>
-                <span class="text-2xl font-semibold text-gray-900">{{ formatCurrency(attributionTotals().firstTouchRevenue) }}</span>
-              </div>
-            </lfx-card>
-            <lfx-card styleClass="flex-1 min-w-[100px]">
-              <div class="flex flex-col gap-1">
-                <span class="text-sm text-gray-500">Last Touch</span>
-                <span class="text-2xl font-semibold text-gray-900">{{ formatCurrency(attributionTotals().lastTouchRevenue) }}</span>
-              </div>
-            </lfx-card>
-            <lfx-card styleClass="flex-1 min-w-[100px]">
-              <div class="flex flex-col gap-1">
-                <span class="text-sm text-gray-500">Linear</span>
-                <span class="text-2xl font-semibold text-gray-900">{{ formatCurrency(attributionTotals().linearRevenue) }}</span>
-              </div>
-            </lfx-card>
-            <lfx-card styleClass="flex-1 min-w-[100px]">
-              <div class="flex flex-col gap-1">
-                <span class="text-sm text-gray-500">Time Decay</span>
-                <span class="text-2xl font-semibold text-gray-900">{{ formatCurrency(attributionTotals().timeDecayRevenue) }}</span>
-              </div>
-            </lfx-card>
-          </div>
-
-          <!-- Channel Attribution Table -->
-          <div class="flex flex-col gap-4 p-4 border border-gray-200 rounded-lg" data-testid="email-ctr-drawer-attribution-table">
-            <div class="flex flex-col gap-1">
-              <h3 class="text-sm font-semibold text-gray-900">Revenue by Channel</h3>
-              <p class="text-sm text-gray-600">Multi-touch attribution across marketing channels over the last 6 months</p>
-            </div>
-
-            <!-- Table Header -->
-            <div class="flex items-center py-2 px-1 text-[10px] font-semibold text-gray-500 uppercase tracking-wider border-b border-gray-200">
-              <span class="flex-[2] min-w-[120px]">Channel</span>
-              <span class="flex-1 text-right">Sessions</span>
-              <span class="flex-1 text-right">First Touch</span>
-              <span class="flex-1 text-right">Last Touch</span>
-              <span class="flex-1 text-right">Linear</span>
-              <span class="flex-1 text-right">Time Decay</span>
-              <span class="w-20 text-right">Rev/Session</span>
-            </div>
-
-            <!-- Channel Rows -->
-            @for (channel of attributionData().channels; track channel.channel) {
-              <div
-                class="flex items-center py-2.5 px-1 border-b border-gray-100 cursor-pointer hover:bg-gray-50 transition-colors"
-                role="button"
-                tabindex="0"
-                [attr.aria-expanded]="expandedChannels().has(channel.channel)"
-                (click)="toggleChannel(channel.channel)"
-                (keydown.enter)="toggleChannel(channel.channel)"
-                (keydown.space)="$event.preventDefault(); toggleChannel(channel.channel)"
-                data-testid="email-ctr-drawer-attribution-channel-row">
-                <span class="flex-[2] min-w-[120px] text-sm font-semibold text-gray-900 truncate pr-2">{{ channel.channel }}</span>
-                <span class="flex-1 text-right text-sm font-semibold text-gray-900">{{ formatCompact(channel.sessions) }}</span>
-                <span class="flex-1 text-right text-sm font-semibold text-gray-900">{{ formatCurrency(channel.firstTouchRevenue) }}</span>
-                <span class="flex-1 text-right text-sm font-semibold text-gray-900">{{ formatCurrency(channel.lastTouchRevenue) }}</span>
-                <span class="flex-1 text-right text-sm font-semibold text-gray-900">{{ formatCurrency(channel.linearRevenue) }}</span>
-                <span class="flex-1 text-right text-sm font-semibold text-gray-900">{{ formatCurrency(channel.timeDecayRevenue) }}</span>
-                <span class="w-20 text-right text-sm font-semibold text-gray-900">{{ revPerSession(channel) }}</span>
-              </div>
-
-              <!-- Expanded Project Rows -->
-              @if (expandedChannels().has(channel.channel)) {
-                @for (project of attributionProjectsByChannel().get(channel.channel) ?? []; track project.projectName) {
-                  <div class="flex items-center py-2 px-1 pl-6 border-b border-gray-50 bg-gray-50" data-testid="email-ctr-drawer-attribution-project-row">
-                    <span class="flex-[2] min-w-[120px] text-xs text-gray-500 truncate pr-2" [title]="project.projectName">{{ project.projectName }}</span>
-                    <span class="flex-1 text-right text-xs text-gray-600">{{ formatCompact(project.sessions) }}</span>
-                    <span class="flex-1 text-right text-xs text-gray-600">{{ formatCurrency(project.firstTouchRevenue) }}</span>
-                    <span class="flex-1 text-right text-xs text-gray-600">{{ formatCurrency(project.lastTouchRevenue) }}</span>
-                    <span class="flex-1 text-right text-xs text-gray-600">{{ formatCurrency(project.linearRevenue) }}</span>
-                    <span class="flex-1 text-right text-xs text-gray-600">{{ formatCurrency(project.timeDecayRevenue) }}</span>
-                    <span class="w-20"></span>
-                  </div>
-                }
-              }
-            }
-
-            <!-- Total Row -->
-            <div class="flex items-center py-2.5 px-1 border-t-2 border-gray-300 font-bold">
-              <span class="flex-[2] min-w-[120px] text-sm text-gray-900">TOTAL</span>
-              <span class="flex-1 text-right text-sm text-gray-900">{{ formatCompact(attributionTotals().sessions) }}</span>
-              <span class="flex-1 text-right text-sm text-gray-900">{{ formatCurrency(attributionTotals().firstTouchRevenue) }}</span>
-              <span class="flex-1 text-right text-sm text-gray-900">{{ formatCurrency(attributionTotals().lastTouchRevenue) }}</span>
-              <span class="flex-1 text-right text-sm text-gray-900">{{ formatCurrency(attributionTotals().linearRevenue) }}</span>
-              <span class="flex-1 text-right text-sm text-gray-900">{{ formatCurrency(attributionTotals().timeDecayRevenue) }}</span>
-              <span class="w-20"></span>
-            </div>
-          </div>
-        </div>
-
-        <!-- Section Divider -->
-        <hr class="border-gray-200" />
-      }
-
       <div class="flex flex-col gap-5" data-testid="email-ctr-drawer-email-section">
         <div class="flex items-center gap-2">
           <i class="fa-light fa-envelope text-blue-500"></i>
